Move graph visiting utilities to a separate file.
